root: Move general functions to util.go
Move some general functions from root.go to util.go.
