+++ cmd/util.go
@@ -2,10 +2,13 @@
 package cmd
 
 import (
+	"fmt"
 	"log"
+	"os"
 	"strconv"
 	"strings"
 
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	flag "github.com/spf13/pflag"
 	"github.com/spf13/viper"
@@ -94,6 +97,136 @@ func getCurrentBranchMR(rn string) int {
 // .git/lab/lab.toml
 func getMainConfig() *viper.Viper {
 	return config.MainConfig
+}
+
+// parseArgs is used by commands to parse command line arguments.
+// This function returns a remote name and number.
+func parseArgs(args []string) (string, int64, error) {
+	if !git.InsideGitRepo() {
+		return "", 0, nil
+	}
+
+	remote, num, err := parseArgsStr(args)
+	if err != nil {
+		return "", 0, err
+	}
+	ok, err := git.IsRemote(remote)
+	if err != nil {
+		return "", 0, err
+	} else if !ok && remote != "" {
+		switch len(args) {
+		case 1:
+			return "", 0, errors.Errorf("%s is not a valid remote or number", args[0])
+		default:
+			return "", 0, errors.Errorf("%s is not a valid remote", args[0])
+		}
+	}
+	if remote == "" {
+		remote = forkedFromRemote
+	}
+	rn, err := git.PathWithNameSpace(remote)
+	if err != nil {
+		return "", 0, err
+	}
+	return rn, num, nil
+}
+
+// parseArgsRemoteString is used by commands to parse command line arguments.
+// This function returns a remote name and the project name.  If no
+// remote name is given, the function returns "" and the project name of
+// the default remote (ie 'origin').
+func parseArgsRemoteString(args []string) (string, string, error) {
+	if !git.InsideGitRepo() {
+		return "", "", nil
+	}
+
+	remote, str := forkedFromRemote, ""
+
+	if len(args) == 1 {
+		ok, err := git.IsRemote(args[0])
+		if err != nil {
+			return "", "", err
+		}
+		if ok {
+			remote = args[0]
+		} else {
+			str = args[0]
+		}
+	} else if len(args) > 1 {
+		remote, str = args[0], args[1]
+	}
+
+	ok, err := git.IsRemote(remote)
+	if err != nil {
+		return "", "", err
+	}
+	if !ok {
+		return "", "", errors.Errorf("%s is not a valid remote", remote)
+	}
+
+	remote, err = git.PathWithNameSpace(remote)
+	if err != nil {
+		return "", "", err
+	}
+	return remote, str, nil
+}
+
+// parseArgsStr is used by commands to parse command line arguments.
+// This function returns a string and number.
+func parseArgsStr(args []string) (string, int64, error) {
+	if len(args) == 2 {
+		n, err := strconv.ParseInt(args[1], 0, 64)
+		if err != nil {
+			return args[0], 0, err
+		}
+		return args[0], n, nil
+	}
+	if len(args) == 1 {
+		n, err := strconv.ParseInt(args[0], 0, 64)
+		if err != nil {
+			return args[0], 0, nil
+		}
+		return "", n, nil
+	}
+	return "", 0, nil
+}
+
+// parseArgsWithGitBranchMR returns a remote name and a number if parsed.
+// If no number is specified, the MR id associated with the current
+// branch is returned.
+func parseArgsWithGitBranchMR(args []string) (string, int64, error) {
+	s, i, err := parseArgs(args)
+	if i == 0 {
+		i = int64(getCurrentBranchMR(s))
+		if i == 0 {
+			fmt.Println("Error: Cannot determine MR id.")
+			os.Exit(1)
+		}
+	}
+	return s, i, err
+}
+
+// setCommandPrefix returns a concatenated value of some of the commandline.
+// For example, 'lab mr show' would return 'mr_show.', and 'lab issue list'
+// would return 'issue_list.'
+func setCommandPrefix(scmd *cobra.Command) {
+	for _, command := range RootCmd.Commands() {
+		if CommandPrefix != "" {
+			break
+		}
+		commandName := strings.Split(command.Use, " ")[0]
+		if scmd == command {
+			CommandPrefix = commandName + "."
+			break
+		}
+		for _, subcommand := range command.Commands() {
+			subCommandName := commandName + "_" + strings.Split(subcommand.Use, " ")[0]
+			if scmd == subcommand {
+				CommandPrefix = subCommandName + "."
+				break
+			}
+		}
+	}
 }
 
 // textToMarkdown converts text with markdown friendly line breaks
